Refactor C code in a separate file
